Setup

+++ setup.py
@@ -4,7 +4,8 @@ import json
 import urllib.request
 import gdown
 import argparse
-
+import platform
+import helpers
 
 ###### common setup utils ##############
 
@@ -35,9 +36,10 @@ def get_conda_envs():
 
 
 def setup_reid(root):
+    print(f'setup_reid(): ======================= start =======================')
     env_name  = cfg.reid_env
     repo_name = "centroids-reid"
-    src_url   = "https://github.com/mikwieczorek/centroids-reid.git"
+    src_url   = "https://github.com/m-iDev-0792/centroids-reid.git"
     rep_path  = "./reid"
 
     if not repo_name in os.listdir(rep_path):
@@ -46,28 +48,41 @@ def setup_reid(root):
 
         # create the models folder inside repo, weights will be added to that folder later on
         models_folder_path = os.path.join(rep_path, repo_name, "models")
-        os.system(f"mkdir {models_folder_path}")
+        try_num = 0
+        while try_num < 10:
+            if not os.path.exists(models_folder_path):
+                print(f'{models_folder_path} does not exist. Create it now. pwd is {os.getcwd()}')
+                os.mkdir(f"{models_folder_path}")
+                try_num += 1
+            else:
+                break            
 
         url = "https://drive.google.com/uc?export=download&id=1w9yzdP_5oJppGIM4gs3cETyLujanoHK8&confirm=t&uuid=fed3cb8a-1fad-40bd-8922-c41ededc93ae&at=ALgDtsxiC0WTza4g47gqC5VPyWg4:1679009047787"
         save_path = os.path.join(models_folder_path, "dukemtmcreid_resnet50_256_128_epoch_120.ckpt")
-        urllib.request.urlretrieve(url, save_path)
+        if not os.path.exists(save_path):
+            print(f'{save_path} does not exist, downloading from {url}')
+            urllib.request.urlretrieve(url, save_path)
 
         url = "https://drive.google.com/uc?export=download&id=1ZFywKEytpyNocUQd2APh2XqTe8X0HMom&confirm=t&uuid=450bb8b7-b3d0-4465-b0c9-bb6f066b205e&at=ALgDtswylGfYgY71u8ZmWx4CfhJX:1679008688985"
         save_path = os.path.join(models_folder_path, "market1501_resnet50_256_128_epoch_120.ckpt")
-        urllib.request.urlretrieve(url, save_path)
+        if not os.path.exists(save_path):
+            print(f'{save_path} does not exist, downloading from {url}')
+            urllib.request.urlretrieve(url, save_path)
 
     if not env_name in get_conda_envs():
-        make_conda_env(env_name, libs="python=3.8")
+        make_conda_env(env_name, libs="python=3.9")
         cwd = os.getcwd()
         os.chdir(os.path.join(rep_path, repo_name))
         os.system(f"conda run --live-stream -n {env_name} conda install --name {env_name} pip")
         os.system(f"conda run --live-stream -n {env_name} pip install -r requirements.txt")
 
         os.chdir(cwd)
+    print(f'setup_reid(): ======================= end =======================')
 
 # clone and install vitpose
 # download the model
 def setup_pose(root):
+    print(f'setup_pose(): ======================= start =======================')
     env_name  = cfg.pose_env
     repo_name = "ViTPose"
     src_url   = "https://github.com/ViTAE-Transformer/ViTPose.git"
@@ -88,12 +103,15 @@ def setup_pose(root):
 
         os.chdir(os.path.join(root, rep_path, "ViTPose"))
         os.system(f"conda run --live-stream -n {env_name} pip install -v -e .")
+        os.system(f"conda run --live-stream -n {env_name} pip install tqdm")
         os.system(f"conda run --live-stream -n {env_name} pip install timm==0.4.9 einops")
-
+        os.system(f'conda run --live-stream -n {env_name} pip install torch==1.9.0+cu111 torchvision==0.10.0+cu111 torchaudio==0.9.0 -f https://download.pytorch.org/whl/torch_stable.html')
+    print(f'setup_pose(): ======================= end =======================')
 
 # clone and install str
 # download the model
 def setup_str(root):
+    print(f'setup_str(): ======================= start =======================')
     env_name  = cfg.str_env
     repo_name = "parseq"
     src_url   = "https://github.com/baudm/parseq.git"
@@ -108,25 +126,43 @@ def setup_str(root):
 
     if not env_name in get_conda_envs():
         make_conda_env(env_name, libs="python=3.9")
-        os.system(f"make torch-cu117")
+        cuda_ver = 'cu117'
+        if platform.system() == 'Windows':
+            pwd = os.getcwd()
+            print(f'Copy {pwd}/../requirements to {pwd}/requirements')
+            # Windows does not have a shell terminal, so just copy prebuilt requirements
+            helpers.copy_folder_contents("../requirements", "requirements")
+        else:
+            os.system(f"make torch-{cuda_ver}")
         os.system(f"conda run --live-stream -n {env_name} conda install --name {env_name} pip")
-        os.system(f"conda run --live-stream -n {env_name} pip install -r requirements/core.cu117.txt -e .[train,test]")
+        os.system(f"conda run --live-stream -n {env_name} pip install -r requirements/core.{cuda_ver}.txt -e .[train,test]")
 
     os.chdir(root)
+    print(f'setup_str(): ======================= end =======================')
 
 def download_models_common(root_dir):
+    print(f'download_models_common(): ======================= start =======================')
     repo_name = "ViTPose"
-    rep_path = "./pose"
+    rep_path = "pose"
 
     url = cfg.dataset['SoccerNet']['pose_model_url']
-    models_folder_path = os.path.join(rep_path, repo_name, "checkpoints")
-    if not os.path.exists(models_folder_path):
-        os.system(f"mkdir {models_folder_path}")
-    save_path = os.path.join(rep_path, "ViTPose", "checkpoints", "vitpose-h.pth")
+    models_folder_path = os.path.join(root_dir, rep_path, repo_name, "checkpoints")
+    try_num = 0
+    while try_num < 10:
+        if not os.path.exists(models_folder_path):
+            print(f'{models_folder_path} does not exist. Create it now. pwd is {os.getcwd()}')
+            os.mkdir(f"{models_folder_path}")
+            try_num += 1
+        else:
+            break
+    save_path = os.path.join(root_dir, rep_path, "ViTPose", "checkpoints", "vitpose-h.pth")
     if not os.path.isfile(save_path):
+        print(f'Downloading model from {url}')
         gdown.download(url, save_path)
+    print(f'download_models_common(): ======================= end =======================')
 
 def download_models(root_dir, dataset):
+    print(f'download_models(): ======================= start =======================')
     # download and save fine-tuned model
     save_path = os.path.join(root_dir, cfg.dataset[dataset]['str_model'])
     if not os.path.isfile(save_path):
@@ -137,8 +173,10 @@ def download_models(root_dir, dataset):
     if not os.path.isfile(save_path):
         source_url = cfg.dataset[dataset]['legibility_model_url']
         gdown.download(source_url, save_path)
+    print(f'download_models(): ======================= end =======================')
 
 def setup_sam(root_dir):
+    print(f'setup_sam(): ======================= start =======================')
     os.chdir(root_dir)
     repo_name = 'sam2'
     src_url = 'https://github.com/davda54/sam'
@@ -146,6 +184,7 @@ def setup_sam(root_dir):
     if not repo_name in os.listdir(root_dir):
         # clone source repo
         os.system(f"git clone --recurse-submodules {src_url} {os.path.join(root_dir, repo_name)}")
+    print(f'setup_sam(): ======================= end =======================')
 
 
 if __name__ == '__main__':
